Setting the default date to today is a bad idea (#5332)
If a user selects a year, the day and month are set to today's day and month. This is a very bad idea since people want to sort entries by date. Most of the time only the year is of interest. The entries will be sorted then by the time and day when the were created...
This commit should set the values to the same by default.