Fix O2M update permission checking (#23915)
* Fetch permissions corresponding to the validated action

* Fetch permission flags for O2M relations

* Only validate access against the primary key field if we can reasonably assume that the user want's to update it

* Filter out `null` selects

* Fix unit test

* Respect action for fields check

Before this, "read" permissions were used when checking permissions for
"updates".

* Adjust options for parseFields in import-export.ts

* Fix comment typo

* Properly fix issue with update requiring pk perms

While in commit 918de2f we ensured the correct perms ("update" perms
instead of "read" perms) are checked for rel item updates, now the update
permission on the primary key field was still required since the primary
key has been part of the payload. This then failed a bit later in the
payload field check performed in the `processPayload` function.
So the proper fix here is to omit the primary key from the payload for
updates (primary key cannot change anyway). With this we can also drop
the fields filter workaround in the `updateMany` method.

Tested with all relationship types.

* Construct the AST ourselves, as we are only interested in all the top level fields of an item, regardless if they relate to any actual DB backed field

* Fix test

* Try to fix import issue in tests

* Adjust recently added test cases

* Clean-up

Basically reverting changes from 918de2f (respecting action) as they are
no longer required with 2290444 (constructing own AST for validating
item access)

* Inline children to make AST construction clearer

* Remove o2mNodes from fetchPermittedAstRootFields as it is no longer needed

* Remove unused import

* Revert changes in `getDBQuery` since they aren't strictly necessary anymore, since every node in the AST is a `FieldNode` now

---------

Co-authored-by: Pascal Jufer <pascal-jufer@bluewin.ch>